requirements.txt
py-clob-client>=0.0.1
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.7.0
flask>=3.0.0
scipy>=1.12.0
numpy>=1.26.0
